Eliminate ./ from paths in tar (filesystem) (#77)
When creating the tar file for the filesystem, it should not have ./
before the filenames. If it does, then the readfile shell command
implemented in shell.c doesn't work, because it expects the filename
without the ./

run.sh in github is like this, so it only needs to be updated in the
website.

Thanks a lot for the book, it is really nice!