Fix EN explanation of manually reading memory mapping
Previously, the explanation said that the second-level table for
0x80000000 will have its first 512 entries 0. This is incorrect,
because VPN[0] of 0x80200000 is 0, so the second-level table should
have non-zero entries from its beginning. I assume the explanation was
intending to discuss the first-level table, since VPN[1] of
`__kernel_base` is 0x200.

Due to this edit, the explanation would end after finding the
second-level page table without actually verifying anything about
it. I therefore continued by inspecting the contents of the
second-level page table to verify that the mapped physical address
corresponds to the virtual address you started with. In order to do
this, I had to change the virtual address we were investigating from
0x80000000 to 0x80200000, because the former is before `__kernel_base`
and so does not get an entry in the page table.