Fix explanation of manually reading memory mapping (#90)
Previously, the explanation said that the second-level table for
0x80000000 will have its first 512 entries equal to 0. This is
incorrect, because VPN[0] of 0x80200000 is 0, so the second-level table
should have non-zero entries from its beginning. I assume the
explanation was intending to discuss the first-level table, since VPN[1]
of `__kernel_base` is 0x200. (I've verified this in qemu, though of
course it's possible there's a bug in my code somewhere and that I've
misunderstood the specification!)

Due to this edit, the explanation would end after finding the
second-level page table without actually verifying anything about it. I
therefore continued by inspecting the contents of the second-level page
table to verify that the mapped physical address corresponds to the
virtual address you started with. In order to do this, I had to change
the virtual address we were investigating from 0x80000000 to 0x80200000,
because the former is before `__kernel_base` and so does not get an
entry in the page table.

I can't speak any of the other languages that the website is translated
into, so would need someone else to do the translations, sorry.